+++ batch_scoring.py
@@ -1,4 +1,6 @@
+import sys
+
 from datarobot_batch_scoring import main
 # This exists as a simple entrypoint for PyInstaller
 
-main.main()
+sys.exit(main.main())